Tela Criar Alimento

+++ app/src/main/java/entities/Food.java
@@ -10,16 +10,16 @@ public class Food {
     @DocumentId
     private String documentId;
     public String name;
-    public int protein;
+    public double protein;
     public int calories;
-    public int carb;
-    public int fat;
-    public int fiber;
+    public double carb;
+    public double fat;
+    public double fiber;
     public int quantity;
     public boolean unity;
 
 
-    public Food(String name, int protein, int calories, int carb, int fat, int fiber, int quantity, boolean unity) {
+    public Food(String name, double protein, int calories, double carb, double fat, double fiber, int quantity, boolean unity) {
         this.name = name;
         this.protein = protein;
         this.calories = calories;
